Add Car and ServiceItem classes with sample data
Introduces `Car` and `ServiceItem` classes in the
`CarService.Web.Models` namespace. The `Car` class
includes properties for vehicle details and a list
of service items. The `ServiceItem` class defines
maintenance tasks with relevant intervals. A new
`CarService` class is added to manage a list of
cars, initializing a sample car with predefined
service items and their details.